Make indirection nodes a single word.
By stealing one more bit of the FUN pointer in a node we can
use the FUN pointer as the indirection pointer.  This means
one less load to follow an indirection and one less store to create
an indirection.  That means that a single (atomic) store can be
used to update a node without the need for a lock, if we ever
get multi-processor support.