Add a horrible hack to make some uses of lenses work.
MicroHs has OverloadedRecordDot permanently on, this means
that a.b is interpreted as record selection.  But when using
lenses it should be interprteed as function composition.
This hack looks at b, and if it looks more like a function than
a label, then a.b is translated into 'a . b', i.e., composition.
This sometimes works, but not in general since parsing has already
messed things up.  if you write 'f a.b' then with OverloadedRecordDot
it is parsed as 'f (a.b)', whereas with composition it would be '(f a). b'.